Implemented user-driven placement of reinforcements
While there are still reinforcements to place, will display contents of
the contriesOwned vector of the player and
will ask for a country name.
Will check to see if this country name is in the vector. If found will
ask for reinforcements to add.
Will value is valid (not more than reinforcements allowed) will add this
value to current armies existing on the country.