error message bugs fixed
fixed bug in fortification phase where multiple error messages were
displayed if an invalid input was made instead of just one.

fixed bug were upon a successful reinforcement would not display that
the reinforcement  was a success before refreshing the map and clearing
the screen.

fixed a few couts to be spaced nicer